Rename and reduce system
