+++ openpathsampling/experimental/simstore/class_info.py
