+++ openpathsampling/experimental/simstore/class_lookup.py
