+++ openpathsampling/experimental/simstore/custom_json.py
