+++ openpathsampling/experimental/simstore/my_types.py
