+++ openpathsampling/experimental/simstore/serialization.py
