+++ openpathsampling/experimental/simstore/serialization_helpers.py
