+++ openpathsampling/experimental/simstore/test_class_info.py
