+++ openpathsampling/experimental/simstore/test_custom_json.py
