+++ openpathsampling/experimental/simstore/test_serialization.py
