+++ openpathsampling/experimental/simstore/test_serialization_helpers.py
