+++ openpathsampling/experimental/simstore/tools.py
