+++ openpathsampling/experimental/storage/__init__.py
@@ -1,13 +1,8 @@
 # TODO: improve the imports here
-from . import backend
-from . import class_info
-from . import my_types
-from . import ops_storage
-from . import serialization
-from . import serialization_helpers
 from . import snapshots
-from . import sql_backend
-from . import storage
-from . import tools
 from . import monkey_patches
 from . import dict_serialization_helpers
+
+from .ops_storage import (
+    Storage, OPSClassInfoContainer, ops_schema, ops_class_info
+)
