+++ openpathsampling/experimental/storage/ops_storage.py
@@ -1,30 +1,30 @@
-from . import storage
-from . import sql_backend
+from ..simstore import storage
+from ..simstore import sql_backend
 
-from .serialization_helpers import to_json_obj as json_serializer
-from .serialization_helpers import from_json_obj as deserialize_sim
-from .serialization_helpers import import_class
-from .serialization_helpers import get_uuid, set_uuid
-from .serialization_helpers import default_find_uuids
+from ..simstore.serialization_helpers import to_json_obj as json_serializer
+from ..simstore.serialization_helpers import from_json_obj as deserialize_sim
+from ..simstore.serialization_helpers import import_class
+from ..simstore.serialization_helpers import get_uuid, set_uuid
+from ..simstore.serialization_helpers import default_find_uuids
 
-from .class_lookup import ClassIsSomething
+from ..simstore.class_lookup import ClassIsSomething
 
 import openpathsampling as paths
 from openpathsampling.netcdfplus import StorableObject
 
-from . import tools
+from ..simstore import tools
 
-from .custom_json import (
+from ..simstore.custom_json import (
     JSONSerializerDeserializer,
     numpy_codec, bytes_codec, uuid_object_codec,
 )
 
-from .serialization import (
+from ..simstore.serialization import (
     ToDictSerializer, SchemaSerializer, SchemaDeserializer,
     SimulationObjectSerializer
 )
 
-from .class_info import ClassInfo, ClassInfoContainer
+from ..simstore.class_info import ClassInfo, ClassInfoContainer
 
 from . import snapshots
 import logging
@@ -190,12 +190,12 @@ ops_simulation_classes = {
 }  # TODO: add more to these
 
 
-class OPSStorage(storage.GeneralStorage):
+class Storage(storage.GeneralStorage):
     def __init__(self, backend, schema, class_info, fallbacks=None,
                  safemode=False):
         # TODO: this will change to match the current notation
-        super(OPSStorage, self).__init__(backend, schema, class_info,
-                                         fallbacks)
+        super(Storage, self).__init__(backend, schema, class_info,
+                                      fallbacks)
 
         self.n_snapshot_types = 0
 
@@ -212,7 +212,7 @@ class OPSStorage(storage.GeneralStorage):
         class_info = tools.none_to_default(class_info, ops_class_info)
         simulation_classes = tools.none_to_default(simulation_classes,
                                                    ops_simulation_classes)
-        super(OPSStorage, obj).__init__(
+        super(Storage, obj).__init__(
             backend=backend,
             schema=schema,
             class_info=class_info,
