+++ openpathsampling/experimental/storage/snapshots.py
@@ -1,7 +1,7 @@
 # NOTE: this is part of the OPS-specific stuff
-from .class_info import ClassInfo
+from ..simstore.class_info import ClassInfo
 
-from .serialization_helpers import get_uuid
+from ..simstore.serialization_helpers import get_uuid
 
 def _nested_schema_entries(schema_entries, lazies):
     """Recursive algorithm to create all schema entries
